Recipe Popup created. TODO: manage open and close and load all data
